gb_count/fix coupons
